pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cis-to-fleet"
version = "0.1.0"
description = "Convert CIS benchmarks to Fleet-compatible policy files"
readme = "README.md"
requires-python = ">=3.9,<4.0"
license = {text = "MIT"}
authors = [
    {name = "cis-to-fleet contributors"},
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: System :: Systems Administration",
    "Topic :: Security",
]

dependencies = [
    "requests>=2.31.0",
    "typer>=0.12.0",
    "ruamel.yaml>=0.18.0",
    "textual>=0.59.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-httpx>=0.21.0",
    "responses>=0.23.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
    "build>=0.10.0",
]

[project.scripts]
cis-to-fleet = "cis_to_fleet.__main__:app"

[project.urls]
Homepage = "https://github.com/username/cis-to-fleet"
Repository = "https://github.com/username/cis-to-fleet"
Issues = "https://github.com/username/cis-to-fleet/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"